Use double for max_ranking too, remove first half base-points distribution (done elsewhere)

+++ src/network/protocols/server_lobby.hpp
@@ -80,13 +80,14 @@ private:
 
     /* Ranking related variables */
 
+    // If updating the base points, update the base points distribution in DB
     const float BASE_RANKING_POINTS   = 4000.0f;
     const float MAX_SCALING_TIME      = 600.0f;
     const float MAX_POINTS_PER_SECOND = 0.125f;
 
     std::vector<double>       m_rankings; // TODO : convert from and to int when communicating with the server. Think to round it correctly
     std::vector<unsigned int> m_num_ranked_races;
-    std::vector<int>          m_max_ranking;
+    std::vector<double>       m_max_ranking;
 
     // connection management
     void clientDisconnected(Event* event);
